Disable Ctrl-S and Ctrl-Q
By default, Ctrl-S and Ctrl-Q are used for software flow control.
Ctrl-S stops data from being transmitted to the terminal until
you press Ctrl-Q. This originates in the days when you might want to
pause the transmission of data to let a device like a printer catch up.
Need to turn this feature off in order to detect keyboard event Ctrl-Q