Fix output sliding right
In raw mode, input is not processed as it is in cooked mode, meaning that no
automatic newline conversion or terminal processing is performed. As a result,
when we printf somthing (e.g."Ctrl+P detected!\n"), the newline character '\n'
moves the cursor to the next line, but does not return the cursor to the beginning of
the line (carriage return), leading to subsequent lines being printed to the
right.

So we disabled raw mode whenever there is no keyboard event during each
corutine and enabled the raw mode before the keyboard detection