Fix stack overflow when running PC vs PC again
It's werid that tha value of i should be reset back to 0 when
corutine_ai() was running again. However, it seems like the value of i
ramined at 2 which was the value of the last time corutine_ai() was
ended.
Therefore, we need to reset the value of i back to what it was before
each time corutine_ai() ended.